:sparkles: Complete MainExporter to Record Results
The main changes in this commit focus on the `MainExporter` class.

We have added two new dependencies in the class constructor, `notFoundListItemWriter` and `foundListItemWriter`, which are both instances of `ItemRecorder` for `NicoListItem` and `VocaDBSongListItem` respectively. This indicates a shift towards a more modular architecture where each class has a single responsibility.

We have also modified the `run` method in `MainExporter`. We now record all the songs that were either found or not found. The recording process is done by the newly added `notFoundListItemWriter` and `foundListItemWriter`.

The logger messages have also been updated to reflect these changes. The logger message now includes additional information about the recording process of the results. The logging for the recording process is done right after the filtering of the results, and an additional logging message is added after the recording process is done.

Overall, these changes should make the `MainExporter` class more maintainable and its behaviour clearer to understand.