:recycle: Refactored Cache Interface and Related Classes
This commit refactors the cache interface and the classes that interact with it. The PvToVocaDbIdCache interface has been renamed to KVCache, and its method signatures have been updated accordingly. This refactoring takes a more generic approach which is not tied to any specific type of data that needs to be cached.

The classes InVMPvToVocaDbIdCache, ApiBasedVocaDbPvMapper and PvToVocaDbIdCachePersistor have all been updated to use the new KVCache interface. This change is mainly reflected in the updated import statements and the changed class and method signatures.

The PvToVocaDbIdCachePersistor class has also been updated to improve its handling of null values. Previously, null values were not properly handled when reading from or writing to the cache. This commit ensures that null values are properly handled, by inserting nulls for missing columns and checking for empty strings when reading vocadbId from the cache.

These changes should make the cache interface and related classes more flexible and robust, and easier to use with different types of data.