:sparkles: Add VocaDb Compatible SongList CSV Producer
This commit introduces a new VocaDbCompatibleSongListCsvProducer class in the core logic of the Nico to VocaDB songlist exporter. This class implements the ItemRecorder interface and the Closeable interface. The class is responsible for writing VocaDBSongListItems to a specified path in a CSV format that is compatible with VocaDB.

The VocaDbCompatibleSongListCsvProducer class uses a buffered writer to write each item to the target file. It keeps a counter of the number of items written and increments it with each item. The `record` method writes the item's id and note to the file, separated by commas.

Additionally, updates have been made to the ItemRecorder interface. New comments have been added to the interface to provide more clarity on the option to discard the `record` method if writing one-by-one is not efficient or possible.