:card_file_box: Added *.csv to .gitignore
In this commit, we have made an addition to the .gitignore file in our project. The *.csv files, which are typically used for data storage, are now ignored by Git. This change is crucial as we don't want to include these files in our version control system since they are not part of the source code and can often be quite large. It will keep our repository clean and only focused on the actual code. The files are often user or machine-specific and do not have to be shared among all developers.

In addition, we ensured that there is a newline at the end of the file. This is a common convention to ensure better compatibility across different platforms and tools, and to conform to the POSIX definition of a line.