:sparkles: **Add new configuration files for nico2vocadb-songlist-exporter**
This commit introduces two new configuration files, `application.yml` and `application-test.yml`, for the nico2vocadb-songlist-exporter application.

The `application.yml` file provides the main application configuration, including settings related to input/output, user preferences, system settings, and spring boot configuration. The configuration allows the user to specify things such as the NicoNico list ID to be exported, the output CSV file path, and whether to use a private API or the VocaDB's CSV import feature.

The `application-test.yml` file is a similar configuration file meant to be used for testing purposes. It also covers the areas of input/output, user preferences, and system settings. However, it comes with pre-filled values suitable for testing, such as a specific NicoNico list ID and file paths for output and not-found CSV files.

These new configuration files bring more flexibility and convenience to the application, making it easier for users to customize their usage based on their needs and preferences.