:sparkles: Add main method
This commit introduces two new files `Nico2VocaDbSongListExporter.kt` and `Main.kt` in the `nico2vocadb-songlist-exporter` module.

`Nico2VocaDbSongListExporter.kt` is the entry point of the application marked with `@SpringBootApplication` and `@ConfigurationPropertiesScan`. This class will be responsible for the initialization and configuration of the Spring Boot application.

`Main.kt` includes the `main` function, which is the entry point of the Kotlin application. It calls `runApplication<Nico2VocaDbSongListExporter>(*args)` to start the Spring Boot application.