:sparkles: Added PostConstruct annotation to createFolders method in IOConfigImpl.kt
This commit includes a refactor on the `IOConfigImpl.kt` file in the `nico2vocadb-songlist-exporter` module. The `createFolders()` method, which was previously called in the initialization block, now has a `@PostConstruct` annotation.

The `@PostConstruct` annotation is used to ensure that the method it is annotated with gets executed after dependency injection is done to perform any initialization. This means that the `createFolders()` method will be automatically invoked after the spring bean is initialized.

Also, the `createFolders()` method now overrides a method from its superclass. The `super.createFolders()` call has been added to the method body to maintain the original behavior.