:sparkles: **Add HTTP Configurations for NicoNico and VocaDb**
This commit introduces three new HTTP configuration classes: `VocaDbHttpConfig`, `UserAgentRestClientCustomizer`, and `NicoNicoHttpConfig`.

The `VocaDbHttpConfig` class sets up the HTTP configuration for VocaDb. It creates a `HttpServiceProxyFactory` and a `VocaDbSongByPvApi` bean. `HttpServiceProxyFactory` is configured with a customized `RestClientBuilder` from `VocaDbRestClientCustomizer`.

The `UserAgentRestClientCustomizer` class customizes the `RestClient` with a default user-agent header. The user-agent string is obtained from the `SystemConfig`.

The `NicoNicoHttpConfig` class sets up the HTTP configuration for NicoNico. It creates a `CloseableHttpClient` bean and a `HttpServiceProxyFactory` bean. The `CloseableHttpClient` uses the system properties and optionally a `CookieStore`. `HttpServiceProxyFactory` is configured with a customized `RestClientBuilder` with an Apache HTTP client, `NvApiRestClientCustomizer`, and optionally a `DefaultHeadersRestClientCustomizer` with a user-session cookie header. It also creates a `NicoListApi` bean, and has a `persist` method to save the `CookieStore` upon destruction.