:sparkles: Add CoreComponentSetup.kt
This commit introduces a new Kotlin file, `CoreComponentSetup.kt`, in the `mikufan.cx.vtool.app.n2vex.config` package. This setup file is crucial for the configuration of our application, with a key focus on the Nico to VocaDB songlist exporter component.

The new setup class contains several `@Bean` annotations for Spring configuration, each responsible for setting up a distinct part of the application.

The `nicoListFetcher` function creates a new `ApiBasedNicoListFetcher` with a NicoListApi and a preference for using a private API. The `vocadbPvMapper` is another function that sets up a VocaDB PV mapper with a VocaDbSongByPvApi and a cache for PV to Long mapping.

For recording CSV outputs, two functions are provided: `notFoundCsvRecorder` and `outputCsvRecorder`. The former sets up a CSV recorder for NicoListItems that were not found, while the latter sets up a CSV recorder for output, with a decision based on whether the VocaDB CSV format is preferred.

Finally, the `mainExporter` function sets up the main exporter for the application, integrating all the components configured above as its dependencies. This main exporter includes the IOConfig, preference, nicoListFetcher, vocadbPvMapper, notFoundCsvRecorder, and outputCsvRecorder.

This commit is a significant step in our effort to make the Nico to VocaDB songlist exporter component more structured and maintainable.