:sparkles: Add Nico2VocaDbSongListExporterIT integration test
This commit introduces a new integration test, `Nico2VocaDbSongListExporterIT`, into the `nico2vocadb-songlist-exporter` module. This test is based on the `ShouldSpec` style from Kotest and uses the SpringExtension to integrate with the Spring context.

The test is designed to validate the `MainExporter`'s `run` method under the test profile. This is achieved by asserting that no exceptions are thrown when the main exporter runs. The new integration test should help ensure that the main exporter's runtime behavior is correct and stable, which is crucial for the exporting functionality of the nico2vocadb-songlist-exporter module.