:recycle: Refactored niconico list sort parameters
Changed the sort parameters in the `getPublicList` and `getPrivateList` methods within `NicoListApi.kt` from `String` to `NicoListSortKey` and `NicoListSortOrder` respectively. This was done to enhance type safety and readability.

In `ApiBasedNicoListFetcher.kt`, the `doCallApi` method was updated to reflect these changes. The `NicoListSortKey` and `NicoListSortOrder` objects are now directly passed to the `getPublicList` and `getPrivateList` methods.

In `NicoNicoHttpConfig.kt`, added converters `NicoListSortKeyConverter` and `NicoListSortOrderConverter` to handle the conversion of `NicoListSortKey` and `NicoListSortOrder` objects to `String` respectively. These converters are registered with the `DefaultFormattingConversionService` and passed to `HttpServiceProxyFactory`.