:recycle: Refactor Cache Configuration in Nico2VocaDB Exporter
This commit involves a major refactoring of the cache configuration in the Nico2VocaDB song list exporter module. The main change is the removal of the custom `JacksonJsonNodeMarshaller` and its related configurations. Instead, we are now relying on the default Java serialization provided by Infinispan.

The custom marshaller was previously used to serialize objects to JSON format before storing them in the cache. However, this proved unnecessary and added unnecessary complexity to the cache configuration.

We also removed the `objectMapper` dependency in `EnableCustomPersistentLocationGlobalConfigCustomizer`, a class that customizes the global configuration for the Infinispan cache. This is in line with the removal of the custom JSON marshaller.

Now, the cache configuration is much simpler and easier to understand. The removal of the custom marshaller should not have any impact on the operation of the application, as Infinispan's default Java serialization is more than capable for our use case.