:sparkles: Add build.gradle.kts file and update .gitignore
This commit introduces two major changes. Firstly, it includes the addition of a new build.gradle.kts file in the core/vocadb-songlist-importer directory. This file contains the plugin 'my.kotlin-jvm-lib' and a dependency on the project ':shared:all-model'.

Secondly, the .gitignore file has been updated with an extra line to ignore all files with the .cache extension. This is a proactive measure to prevent unwanted cache files from being tracked by git.