:sparkles: Add new build files for http-api and http-impl components
This commit introduces new build files for two components, `http-api` and `http-impl`. They are both created with the `kotlin-jvm-lib` plugin. The plugin will provide necessary tasks to build, test and publish a Kotlin/JVM library. These two components are part of the project's modular architecture and this change will ensure the correct build process for them.