:recycle: Refactor DefaultHeadersRestClientCustomizer Header Handling
This commit refactors the header handling in `DefaultHeadersRestClientCustomizer`. The `defaultHeaders` parameter is now of type `Map<String, List<String>>`, instead of `MultiValueMap<String, String>`, and the map is converted to a `MultiValueMap` if necessary.

This change simplifies the code in `NicoNicoHttpConfig.kt` by removing the need to manually convert a `Map` to `MultiValueMap` before calling `DefaultHeadersRestClientCustomizer`. Now, the conversion is handled directly in `DefaultHeadersRestClientCustomizer`, if it's needed. This reduces the complexity of the caller code and improves the reusability of `DefaultHeadersRestClientCustomizer`.