:recycle: Refactor HTTP services into separate components
This commit separates HTTP-related services into independent components under 'component/http-service'. Classes related to NicoListApi, VocaDbSongByPvApi, and other HTTP services are moved to this new package.

The component is further divided into 'api' and 'impl', organizing interfaces and their implementations respectively. For example, `NicoListFetcher` (interface) and `NicoListFetcherImpl` (implementation) are placed in 'api' and 'impl' respectively.

The `build.gradle.kts` files are updated to reflect the dependencies of each module after the changes. Some dependencies are removed as they are no longer used in the module, and new dependencies are added for the newly created components.

The commit also includes changes in import statements across various files due to the refactoring.