:recycle: Refactored Dependencies and Validation
This commit addresses changes in two main areas.

Firstly, it refactors the `build.gradle.kts` dependencies for `nico2vocadb-songlist-exporter`. It removes the redundant `spring-boot` and `spring-web` dependencies. This is because `spring-boot-dependencies` already includes these dependencies.

Secondly, it changes the `IOConfig` classes. The `@get:Positive` annotation was moved from `IOConfig.kt` to `IOConfigImpl.kt`. This makes sense because `IOConfig.kt` is an interface, which should be agnostic to validation details. Now, the `IOConfigImpl.kt` class, which is the concrete implementation, is responsible for ensuring the provided `nicoListId` is positive.