:arrow_up_down: Dependency modifications in build files
This commit brings changes to the dependencies in the build files of two modules: `nico2vocadb-songlist-exporter` and `vocadb-songlist-importer`.

In `nico2vocadb-songlist-exporter` module, we have removed the `spring-boot-dependencies` from the API platform. This change has been made to streamline the dependencies of the project and remove unnecessary ones.

On the other hand, we have added `kotlin-jvm-inline-logging` dependency in `vocadb-songlist-importer` module. This additional logging dependency will help in providing more comprehensive logging for the module and enhance the debugging process.

Please note, these changes are intended to optimize the project dependencies and improve the overall project structure.