:truck: Refactored model package structure
The model package structure has been refactored. The 'module' directory has been renamed to 'shared'. This change is reflected in all source files where the 'module' package was imported. This refactoring aims to better represent the shared nature of the models within the application, as they are not tied to a specific module.

For instance, the import statement `import mikufan.cx.vtool.module.model.vocadb.VocaDBSongListItem` has been changed to `import mikufan.cx.vtool.shared.model.vocadb.VocaDBSongListItem` in all relevant files. Similar changes have been made for all other models and their references. This refactoring should not affect the functionality of the application but provides a clearer view of the project structure.