:recycle: Refactor Dependencies and Package Structure
This commit makes significant changes to the dependencies in several build.gradle.kts files and moves the CookieStorePersistor class to a new package.

In the build.gradle.kts files for the io-impl, all-model, http-service-impl, and http-service-api components, the dependency scope for the spring-boot-dependencies and jackson libraries has been changed from 'api' to 'implementation'. This change reduces the visibility of these dependencies to other modules, promoting better encapsulation.

Moreover, the CookieStorePersistor and its implementation, NetscapeTxtCookieStorePersistor, have been moved from the http-service-api component to a new component named apache-httpclient-util-api and apache-httpclient-util-impl respectively. This change reflects a more logical organization of the code based on its functionality, making the codebase easier to navigate.

The new apache-httpclient-util-api component now also depends on the httpclient5 library, which was previously a dependency of the http-service-api component. The http-service-impl component now depends on the new apache-httpclient-util-impl instead of the old http-service-api.

Finally, the spring-boot-configuration-processor has been added as a dependency in spring-app.gradle.kts, and the nico2vocadb-songlist-exporter app now depends on apache-httpclient-util-impl. These changes ensure that the necessary libraries are available where they are needed.