:gear: Update application.yml configuration
This commit adds two additional comments to the `sort-order` configuration in `application.yml` file. The comments clarify under which circumstances the `sort-order` value can be null. Specifically, the `sort-order` value can be null, which implies the default order. However, if the `sort-key` is not null, then the `sort-order` cannot be null.