:recycle: Refactor logging in VocaDbPvMapper
This commit refines the logging mechanism in the `VocaDbPvMapper` class. Previously, the log message "Trying to map VocaDB entry" was always printed regardless of whether the PV was already cached or not.

In the updated version, the log message is now split into two parts. If the PV is already cached, the log message now reads "Found cached VocaDB entry". Alternatively, if the PV is not cached and needs to be fetched from the VocaDB API, the log message will state "Trying to fetch VocaDB entry from VocaDB API". This update provides more accurate and informative log messages that better reflect the process flow.