:recycle: Refactor Configuration Files for Improved Readability and Performance
This commit consists of changes in two configuration files: NicoNicoHttpConfig.kt and CacheConfig.kt.

In NicoNicoHttpConfig.kt, import statements related to CookieStorePersistor and NetscapeTxtCookieStorePersistor have been updated. The original import paths containing 'component.apachehttpclientutil' have been replaced with 'util.apachehttpclient'. This refactoring may lead to better organization and readability of the code.

In CacheConfig.kt, significant changes have been made to improve the overall readability and performance. The function `enableCustomPersistentLocation` now directly receives a directory path as a parameter instead of the entire cacheConfigProperties. Moreover, `enableCachePersistent` has been updated to return an instance of `EnablePersistentCacheManagerConfigurer` with cacheConfigProperties' time to live (ttl) and AllCacheNames() passed as parameters. This improves performance as it reduces the overhead of passing the entire cacheConfigProperties object.

Also, class definitions have been refactored for better performance and readability. In the class `EnableCustomPersistentLocationGlobalConfigCustomizer`, the directory path is now received directly via the constructor, reducing the need for additional processing. Similarly, in the class `EnablePersistentCacheManagerConfigurer`, time to live (ttl) and cache names are now directly received via the constructor, allowing for a more efficient configuration of the cache manager.

These changes aim to improve the readability and performance of the code, making it easier for developers to understand and maintain.